Starting point main
I've started working on my function and the input/output file. Currently
the program takes in a file from the command line and copies the entire
file to a 2D string called string. The Sanders() takes in that string
and parses it for the things our program is required to fix recursively,
that way it fixes nested statements and keeps them in order. It doesn't
function completely yet, but when y'all add your functions it shouldn't
be too hard to get there.